+++ src/main/java/com/huotu/huobanmall/model/app/AppTopConsumeModel.java
@@ -33,7 +33,7 @@ public class AppTopConsumeModel {
     /**
      * 消费额
      */
-    private float money;
+    private double money;
 
     /**
      * 购买单数
@@ -66,11 +66,11 @@ public class AppTopConsumeModel {
 //        this.mobile = mobile;
 //    }
 
-    public float getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(float money) {
+    public void setMoney(double money) {
         this.money = money;
     }
